backend/requirements.txt
fastapi
uvicorn
python-dotenv
snowflake-connector-python
numpy
pydantic